+++ src/main/java/br/com/zalf/prolog/webservice/integracao/transport/SistemaTransportTranslecchi.java
@@ -27,9 +27,10 @@ public final class SistemaTransportTranslecchi extends Sistema {
     @Override
     public void insertModeloChecklist(
             @NotNull final ModeloChecklistInsercao modeloChecklist,
-            @NotNull final DadosChecklistOfflineChangedListener checklistOfflineListener) throws Throwable {
-        throw new BloqueadoIntegracaoException("Devido à integração com o Sistema Transport, " +
-                "a criação de modelos de checklist está bloqueada.");
+            @NotNull final DadosChecklistOfflineChangedListener checklistOfflineListener,
+            final boolean statusAtivo) throws Throwable {
+        // Ignoramos o statusAtivo repassado pois queremos forçar que o modelo de checklist tenha o statusAtivo = false.
+        getIntegradorProLog().insertModeloChecklist(modeloChecklist, checklistOfflineListener, false);
     }
 
     @Override
@@ -38,9 +39,18 @@ public final class SistemaTransportTranslecchi extends Sistema {
             @NotNull final Long codUnidade,
             @NotNull final Long codModelo,
             @NotNull final ModeloChecklistEdicao modeloChecklist,
-            @NotNull final DadosChecklistOfflineChangedListener checklistOfflineListener) throws Throwable {
-        throw new BloqueadoIntegracaoException("Devido à integração com o Sistema Transport, " +
-                "a atualização de modelos de checklist está bloqueada.");
+            @NotNull final DadosChecklistOfflineChangedListener checklistOfflineListener,
+            final boolean sobrescreverPerguntasAlternativas) throws Throwable {
+        // Ignoramos a propriedade sobrescreverPerguntasAlternativas pois queremos que para essa integração todas as
+        // edições de perguntas e alternativas sobrescrevam os valores antigos sem alterar os códigos existentes.
+        getIntegradorProLog()
+                .updateModeloChecklist(
+                        token,
+                        codUnidade,
+                        codModelo,
+                        modeloChecklist,
+                        checklistOfflineListener,
+                        true);
     }
 
     @Override